using load as unique app validator
